Conflicts:
	src/main/resources/META-INF/persistence.xml
	src/main/resources/config/jdbc.properties

+++ src/main/java/com/dc/esb/servicegov/service/impl/StatisticsServiceImpl.java
@@ -129,7 +129,7 @@ public class StatisticsServiceImpl implements StatisticsService{
      */
     public long getServiceInvokeCount(String type){
         String hql = "select count(*) from " + ServiceInvoke.class.getName() + " as si where si.type = ?";
-        long count = serviceInvokeDAO.findUnique(hql, type);
+        long count = (Long)serviceInvokeDAO.findUnique(hql, type);
         return count;
     }
     /**
@@ -137,7 +137,7 @@ public class StatisticsServiceImpl implements StatisticsService{
      */
     public long getServiceInvokeCount(String systemId, String type){
         String hql = "select count(*) from " + ServiceInvoke.class.getName() + " as si where si.systemId = ? and si.type = ?";
-        long count = serviceInvokeDAO.findUnique(hql, systemId, type);
+        long count = (Long)serviceInvokeDAO.findUnique(hql, systemId, type);
         return count;
     }
 
@@ -198,7 +198,7 @@ public class StatisticsServiceImpl implements StatisticsService{
                     hql2 += " and o.optDate < '" + values.get("endDate")[0] + "' ";
                 }
             }
-            long hisNum = operationHisDAO.findUnique(hql2, pk.getServiceId(), pk.getOperationId());
+            long hisNum = (Long)operationHisDAO.findUnique(hql2, pk.getServiceId(), pk.getOperationId());
             operationReleaseNum += hisNum;
             if(!serviceIds.contains(pk.getServiceId()) && hisNum > 0){
                 serviceIds.add(pk.getServiceId());
